Fix pdf collisions script to also work if executed in another directory
If the collisions/scripts/pdf.py script gets executed in another directory
than the scripts directory, some references cannot be resolved. To fix this,
all files that are included in the repository must be referenced relative
to the script folder instead of the working dir.

With this fix it's now possible to execute the script like this:
   ./pocs/collisions/scripts/pdf.py some.pdf another.pdf